report error works correctly (#19)
Across the app I am now collecting the information I need and passing the user to the reporting software, when possible.

This also fixes some types in network requests.

I had to extend the `NextApiRequest` interface, because the initial request is actually be received as a standard `Request` type which includes a `.json()` async method to parse it. Unfortunately, while I'm sure next has great middleware, it is really a problem because the request is only becoming of type NextApiRequest after I've already received the request and I pass it along to a different function.